// src/app/enums/enums.ts
export enum TypeBarcode {
  ALL = 'ALL',
  PALLET = 'PALLET',
  SEAL = 'SEAL',
  LOCATION = 'LOCATION',
  PRODUCT = 'PRODUCT',
  NF = 'NF',
  CONTAINER = 'CONTAINER',
  Printer = 'Printer',
  Local = 'Local',
  Container = 'Container',
  AlternativeProducts = 'AlternativeProducts',
  ProductPresentation = 'ProductPresentation',
  TfaFailure = 'TfaFailure',
}

export enum TypeResult {
  SUCCESS = 'SUCCESS',
  ERROR = 'ERROR',
  WARNING = 'WARNING',
}

export enum ActionType {
  CREATE = 'CREATE',
  EDIT = 'EDIT',
  DELETE = 'DELETE',
}